Added element accessor aliases
These aliases are somewhat experimental. These seem easier to visually
parse, and I think they more clearly convey the meaning of "element
access" as an extraction process. The only concern is that they might
break, or at least weaken, the declarative semantics of the matrix
language that this project seeks to provide. For example, if the symbol
"e12<-3x3" is read as "extract element 12 from the 3x3 matrix", then
this is an imperative statement. However, if the symbol "e12<-3x3" is
taken as the name of a project operator that relates a 3x3 matrix to a
number equal to the 1,2 element, then it is a declarative statement.

Yet another issue is that I am probably over thinking this, and I should
probably move on to more important issues!